[Code] Added <prerequisites.h> with forward decl~s
