+++ source/server/prerequisites.h
@@ -1,0 +1,16 @@
+
+#pragma once
+
+// Forward declarations
+class Sequencer;
+class SWInetSocket;
+class Broadcaster;
+class Receiver;
+class Listener;
+class UserAuth;
+class ScriptEngine;
+
+namespace Http
+{
+    class Response;
+}
